+++ config/environments/production.rb
@@ -238,6 +238,22 @@ Markus::Application.configure do
   config.x.logging.old_files = 10
 
   ###################################################################
+  # Email Notifications
+  ###################################################################
+  config.action_mailer.delivery_method = :smtp
+  config.action_mailer.smtp_settings = {
+      address:              'smtp.gmail.com',
+      port:                 587,
+      domain:               'example.com',
+      user_name:            'example email',
+      password:             'example password',
+      authentication:       'plain',
+      enable_starttls_auto: true
+  }
+  config.action_mailer.asset_host = 'http://localhost:3000'
+  config.action_mailer.default_url_options = {host: 'localhost:3000'}
+
+  ###################################################################
   # Resque queues
   ###################################################################
 
@@ -245,6 +261,8 @@ Markus::Application.configure do
   config.x.queues.create_groups = 'CSC108'
   # The name of the queue where jobs to collect submissions wait to be executed.
   config.x.queues.collect_submissions = 'CSC108'
+  # The name of the queue where jobs to download submissions wait to be executed.
+  config.x.queues.download_submissions = 'CSC108'
   # The name of the queue where jobs to uncollect submissions wait to be executed.
   config.x.queues.uncollect_submissions = 'CSC108'
   # The name of the queue where jobs to update repos with the list of required files wait to be executed.