+++ Changelog.md
@@ -14,6 +14,9 @@
 - Ignore the "Total" column when uploading a csv file to a grade entry form. This makes the upload and download format
   for the csv file consistent (#4425)
 - Added git hook to limit the maximum file size committed and/or pushed to a git repository (#4421)
+- Display newlines properly in flash messages (#4443)
+- Api calls will now return the 'hidden' status of users when accessing user data (#4445)
+- Make bulk submission file downloads a background job (#4463)
 
 ## [v1.8.4]
 - Fixed bug where test output was not being properly hidden from students (#4379)